pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
    "mypy>=1.2.0",
    "types-requests>=2.29.0.0",
]

[project]
name = "pysos"
version = "0.1.0"
description = "Python interface for Species Observation System API"
authors = [
    {name = "Karl Gunnarsson", email = "karl.ia.gunnarsson@gmail.com"},
]
dependencies = [
    "requests>=2.29.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
